. src/o/symbol.js
