+++ src/o/list.js
@@ -1,7 +1,6 @@
+const AbstractList = require("./abstract_list");
 
 const emptyList;
-
-const AbstractList = require("./abstract_list");
 
 // `List` extends `AbstractList` and is
 // the primary object in Bubblescript and
